viewer/requirements.txt
Flask
requests
pdf417gen
Pillow
qrcode
pyotp
